move code around to prep for alternative impl
